partitions code into a module
